display subtotal display total decrease stock quantity

+++ database/migrations/2022_11_22_012954_create_pass_out_items_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->unsignedBigInteger('item_id');
             $table->unsignedBigInteger('pass_out_id');
             $table->unsignedBigInteger('quantity');
-            $table->unsignedBigInteger('subtotal');
+            $table->decimal('subtotal');
 
             $table->foreign('item_id')
                 ->references('id')
